refactor: move stats implementation into dedicated modules
